Get permalink and put it into chat input
