+++ sclack/component/time.py
@@ -1,0 +1,9 @@
+from datetime import datetime
+
+import urwid
+
+
+class Time(urwid.Text):
+    def __init__(self, timestamp):
+        time = datetime.fromtimestamp(float(timestamp)).strftime('%H:%M')
+        super(Time, self).__init__(('datetime', ' {} '.format(time)))
